requirements-test.txt
pytest==3.10.0
pylint==1.9.3
flake8==3.5.0
autopep8==1.0.4

requests==2.12.4
BeautifulSoup==3.2.1
Pillow==5.4.0
